@@ -4,7 +4,7 @@
     <OutputType>Exe</OutputType>
     <TieredCompilation>false</TieredCompilation>
     <IsPackable>false</IsPackable>
-    <TargetFrameworks>net6.0;net8.0</TargetFrameworks>
+    <TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
